build: Fixes warnings on Linux.
Fixes #44746.

This patch fixes warnings found in libqi on Linux on the configurations
listed below.
It does not fix warnings from external headers (such as boost) though.

Tested configurations:
======================
  - desktop - toolchain feed 'linux64-1604'
    * debug / release
      - GCC v5.5.0 from Ubuntu 16.04 default development packages.
      - GCC v7.4.0
      - GCC v8.3.0
      - GCC v9.1.0
  - android:
    * release:
      - ARM    - toolchain feed 'android-arm'
      - ARM64  - toolchain feed 'android-arm64'
      - x86    - toolchain feed 'android-x86_64'
      - x86_64 - toolchain feed 'android-x86'
  - atom (pepper) - toolchain feed 'linux64-atom'

Warnings found:
===============

-Wpragmas:
----------
warning: unknown option after '#pragma GCC diagnostic' kind.

-Wparentheses:
--------------
warning: unnecessary parentheses in declaration of ‘T’.

-Wnoexcept-type:
----------------
warning: mangled name for ‘T’ will change in C++17 because the exception
specification is part of a function type.

-Wdeprecated-declarations:
--------------------------
warning: ‘T’ is deprecated.
Often T is `std::auto_ptr<U>`.

-Wmismatched-tags:
------------------
warning: 'tuple_size' defined as a struct template here but previously
declared as a class template.

-Wmissing-field-initializers:
-----------------------------
warning: missing initializer for member 'T'.

-Wunneeded-internal-declarations:
---------------------------------
warning: function 'f' is not needed and will not be emitted.

-Wunused-function:
------------------
warning: unused function 'T'.

-Wundefined-var-template:
-------------------------
warning: instantiation of variable 'T' required here, but no definition is available.

-Wsubobject-linkage:
--------------------
warning: ‘T’ has a field ‘m’ whose type uses the anonymous namespace.

Reviewed-on: http://gerrit2.aldebaran.lan/41
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>

+++ src/type/signal.cpp
@@ -430,7 +430,7 @@ QI_WARNING_POP()
     QI_ASSERT(_p);
     // Check arity. Does not require to acquire weakLock.
     // We assume the number of children to int will never be bigger than INT_MAX.
-    QI_ASSERT_TRUE(signature().children().size() <= std::numeric_limits<int>::max());
+    QI_ASSERT_TRUE(signature().children().size() <= static_cast<std::size_t>(std::numeric_limits<int>::max()));
     const auto signalArity = qi::numericConvert<int>(signature().children().size());
     auto subscriberArity = -1;
     Signature subscriberSignature = src.signature();
